add cluster health status in cluster CR
